Prefab Variant changes, and Bullet Parent Script
