+++ BangerProjekt/Assets/Prefabs/Enemies/EnemyBulletParent.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 29930a1ab15235c719e8c5a5bc09d9ad
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
