+++ BangerProjekt/Assets/Prefabs/Enemies/RangedEnemy/RangedEnemyBullet Variant.prefab
@@ -1,0 +1,68 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &1247771786039716016
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.031
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.758
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2614648153934232496, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4431318196554620944, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_Name
+      value: RangedEnemyBullet Variant
+      objectReference: {fileID: 0}
+    - target: {fileID: 6915797070413754180, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_Color.g
+      value: 0.9372549
+      objectReference: {fileID: 0}
+    - target: {fileID: 6915797070413754180, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+      propertyPath: m_Color.r
+      value: 1
+      objectReference: {fileID: 0}
+    m_RemovedComponents:
+    - {fileID: -9010003770214310789, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 29930a1ab15235c719e8c5a5bc09d9ad, type: 3}
