+++ BangerProjekt/Assets/Prefabs/Enemies/RangedEnemy/RangedEnemyBullet Variant.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d5e133a4001bffe09b0d52af4d2a21c8
+guid: 7d0bba7ad0f7bea68bd7ba4377878e5e
 PrefabImporter:
   externalObjects: {}
   userData:
