+++ BangerProjekt/Assets/Scripts/CombatScripts/EnemyScripts/EnemyBullet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6072b4ea5359120be8a8dfdd5178a64e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
